+++ example/lib/main.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 
 import 'package:undo/undo.dart';
@@ -11,6 +9,8 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       theme: ThemeData.light(),
+      darkTheme: ThemeData.dark(),
+      debugShowCheckedModeBanner: false,
       home: HomeScreen(),
     );
   }
@@ -22,63 +22,68 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  final _controller = TrackedStreamController<int>(
-    0,
-    onRedo: (val) => print('Redo -> $val'),
-    onUndo: (val) => print('Undo -> $val'),
-  );
+  SimpleStack _controller;
+
+  @override
+  void initState() {
+    _controller = SimpleStack<int>(
+      0,
+      onUpdate: (val) {
+        if (mounted)
+          setState(() {
+            print('New Value -> $val');
+          });
+      },
+    );
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
-    return StreamBuilder<int>(
-        stream: _controller.stream,
-        builder: (context, snapshot) {
-          final count = snapshot.data;
-          return Scaffold(
-            appBar: AppBar(
-              title: Text('Undo/Redo Example'),
+    final count = _controller.state;
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Undo/Redo Example'),
+      ),
+      body: Center(
+        child: Text('Count: $count'),
+      ),
+      bottomNavigationBar: BottomAppBar(
+        child: Row(
+          children: <Widget>[
+            IconButton(
+              icon: Icon(Icons.arrow_back),
+              onPressed: !_controller.canUndo
+                  ? null
+                  : () {
+                      if (mounted)
+                        setState(() {
+                          _controller.undo();
+                        });
+                    },
             ),
-            body: Center(
-              child: Text('Count: $count'),
+            IconButton(
+              icon: Icon(Icons.arrow_forward),
+              onPressed: !_controller.canRedo
+                  ? null
+                  : () {
+                      if (mounted)
+                        setState(() {
+                          _controller.redo();
+                        });
+                    },
             ),
-            bottomNavigationBar: BottomAppBar(
-              child: Row(
-                children: <Widget>[
-                  IconButton(
-                    icon: Icon(Icons.arrow_back),
-                    onPressed: !_controller.canUndo
-                        ? null
-                        : () {
-                            if (mounted)
-                              setState(() {
-                                _controller.undo();
-                              });
-                          },
-                  ),
-                  IconButton(
-                    icon: Icon(Icons.arrow_forward),
-                    onPressed: !_controller.canRedo
-                        ? null
-                        : () {
-                            if (mounted)
-                              setState(() {
-                                _controller.redo();
-                              });
-                          },
-                  ),
-                ],
-              ),
-            ),
-            floatingActionButtonLocation:
-                FloatingActionButtonLocation.endDocked,
-            floatingActionButton: FloatingActionButton(
-              heroTag: ValueKey('add_button'),
-              child: Icon(Icons.add),
-              onPressed: () {
-                _controller.add(count + 1);
-              },
-            ),
-          );
-        });
+          ],
+        ),
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.endDocked,
+      floatingActionButton: FloatingActionButton(
+        heroTag: ValueKey('add_button'),
+        child: Icon(Icons.add),
+        onPressed: () {
+          _controller.modify(count + 1);
+        },
+      ),
+    );
   }
 }