+++ lib/undo.dart
@@ -1,10 +1,4 @@
 library undo;
-
-import 'dart:async';
-import 'dart:collection';
-
-export 'src/tracked_stream_controller.dart';
-export 'src/tracked_value_notifier.dart';
-
-part 'src/undo_stack.dart';
-part 'src/changes.dart';
+export 'src/version.dart';
+export 'src/undo_stack.dart';
+export 'src/simple_stack.dart';