+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.2.0
+
+- [max] -> [limit]
+- Removing Flutter support for use in AngularDart and web applications
+- Updating tests
+- Deprecating [clear] in favor of [clearHistory]
+- Adding Generics support
+- Simplified for adding group or single change
+- Added label option for every change
+- Updating docs
+- Removing [StreamController](https://gist.github.com/rodydavis/229a69fbc9294a58a46155edef677a8f) and [ValueNotifier](https://gist.github.com/rodydavis/8f58b99999928c3be141fb4c986ebece) Implementations in favor of [Bloc PR](https://github.com/felangel/bloc/tree/master/packages/replay_cubit)
+- Adding [packageVersion]
+- Updating Example
+- Adding [SimpleStack]
+- Adding Dart/Html Web Example
+
 ## 1.1.0
 
 - Adding StreamController and ValueNotifier Implementations